# MentoringARELSwagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger MentoringAERL Students Service"
  },
  "host": "localhost:8080",
  "basePath": "/MentoringAERL-CONTROLLER-1.0-SNAPSHOT",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/universities/": {
      "post": {
        "tags": [
          "universities"
        ],
        "summary": "Create new university",
        "description": "Return created university",
        "operationId": "addNewUniversity",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "University information",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddNewUniversityRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/AddNewUniversityResponse"
            }
          },
          "400": {
            "description": "Invalid input. Please check your input and try again later."
          },
          "500": {
            "description": "Some internal server errors has occured. Please try again later."
          }
        }
      },
      "put": {
        "tags": [
          "universities"
        ],
        "summary": "Set universities status",
        "description": "Return statuses with updated universies",
        "operationId": "setUniversityStatus",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Statuses with usiversities",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SetUniversityStatusRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/SetUniversityStatusResponse"
            }
          },
          "400": {
            "description": "Invalid input. Please check your input and try again later."
          },
          "404": {
            "description": "University not found"
          },
          "500": {
            "description": "Some internal server errors has occured. Please try again later."
          }
        }
      }
    },
    "/universities/{id}/students": {
      "get": {
        "tags": [
          "universities"
        ],
        "summary": "Get information about students of the universit by university ID",
        "description": "Return information about students",
        "operationId": "getStudentsOfUniversity",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of university",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/GetStudentsOfUniversityResponse"
            }
          },
          "400": {
            "description": "Invalid input. Please check your input and try again later."
          },
          "404": {
            "description": "Students or university not found"
          },
          "500": {
            "description": "Some internal server errors has occured. Please try again later."
          }
        }
      }
    },
    "/universities/notAssigned": {
      "get": {
        "tags": [
          "universities"
        ],
        "summary": "Returns students that are not assigned to university or assigned to closed one or that is pending government approval",
        "operationId": "getNotAssignedStudents",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/GetNotAssignedStudentsResponse"
            }
          },
          "404": {
            "description": "Students or university not found"
          },
          "500": {
            "description": "Some internal server errors has occured. Please try again later."
          }
        }
      }
    }
  },
  "definitions": {
...
    "AddNewUniversityRequest": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/University"
        }
      }
    },
    "AddNewUniversityResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/University"
        }
      }
    },
    "GetStudentsOfUniversityResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Student"
          }
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "object"
          }
        },
        "warnings": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      }
    },
    "GetNotAssignedStudentsResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "university": {
                "$ref": "#/definitions/University"
              },
              "students": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Student"
                }
              }
            }
          }
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "object"
          }
        },
        "warnings": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      }
    },
    "SetUniversityStatusRequest": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/University"
            }
          }
        }
      }
    },
    "SetUniversityStatusResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/UniversityStatus"
          }
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "object"
          }
        },
        "warnings": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      }
    },
    "UniversityStatus": {
      "type": "string",
      "enum": [
        "OPEN",
        "CLOSED",
        "ON_HOLIDAYS",
        "PENDING_GOVERNMENT_APPROVAL"
      ]
    },
    "University": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "foundationDate": {
          "type": "string",
          "format": "date"
        }
      }
    },
    "Student": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "course": {
          "type": "integer",
          "format": "int32"
        },
        "marks": {
          "type": "object",
          "properties": {
            "MATH": {
              "type": "integer",
              "format": "int32"
            },
            "PHYSICAL_EDUCATION": {
              "type": "integer",
              "format": "int32"
            },
            "PHILOSOPHY": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    }
  }
}